Update docs for find and findMany functions
I wanted to use batchGetItem method using this package, but I was not able to find any function in docs which provided this functionality. When I went into the code, I found that there was already a function using batchGetItem method, but it was not listed in the docs.  
So I wanted to update the docs to make it clear.